+++ CHANGELOG.md
@@ -7,10 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- Prop `filtersProps` to `QasListView`
+- Added new `filtersProps` in `QasListView` for sending to `QasFilter`.
 - Added new component `QasListItems`.
-
-## 2.10.1 - 2021-10-07
 
 ## 2.10.1 - 2021-10-07
 ### Changed