+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new component `QasListItems`.
 
 ### Fixed
-- Adjusting q-input toogle mask value when value is null
+- Adjusting QasInput toogle mask value when value is null
 
 ## 2.10.1 - 2021-10-07