+++ CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.11.0 - 2021-10-19
 
 ### Added
 - Added new `filtersProps` in `QasListView` for sending to `QasFilter`.
 - Added new component `QasListItems`.
 
 ### Fixed
-- Adjusting QasInput toogle mask value when value is null
+- Adjusting `QasInput` toogle mask value when value is null.
 
 ## 2.10.1 - 2021-10-07